separate controller class
